Introduced Offset
This is a lightweight version of Scope that doesn't store a trailer, nor
is capable to append newlines; it only stores an offset depth and can
additionally manage it through promote() and demote() calls. Scope is
implemented through Offset. setOffset(), getOffset() and toString() are
gone due to limited usability; all necessary cases are now covered by a
combination of Offset and Scope