Support body inlining
This is a case when the body schema is a freeform object or refers to a single type (the case of a byte stream). Previously there was no way to find out if a JSON wrapper around the body schema is needed; now it can be bypassed by checking "inlineBody" parameter (which is also the variable contents of which have to be put verbatim into the request body).