Identifiers map + changes in config file organisation
This allows to substitute the identifier used in the API with another name in the generated code, while keeping the API-interacting parts working. Closes #24, which was the main case for implementation of the feature.

The config file now packs up items into two big chapters, "analyzer" and "mustache" (with an immodest idea that eventually "analyzer" might allow options like "openapi" and "raml", and "lua" might emerge next to "mustache" but that's very far ahead).