Mustache: make '_delimiter' work everywhere
The previous implementation of '_delimiter' only applied to the files
it was defined on, making it pretty much unusable since a lot of logic
is defined in 'gtad.yaml', and prepending every definition with the
delimiter override is impractical. The problem was that the main context
defined in 'gtad.yaml' was only passed when rendering files but not
when rendering types and imports. Moreover, Mustache does not pass it
over when traversing to/from partials so even file templates would need
a delimiter override at the beginning of each.

This commit adds GtadContext to all renderers, leaving the partials
problem to the next one. To save on copying things around, GtadContext
is only created anew for types (renderType() recurses and inheriting
the context from the previous level of recursion is undesirable) and
imports reuse the "main" one, push()ing and pop()ping data overlays
as needed. To further optimise and protect from mistakes, km::object
is used instead of type-erasing km::data for the top-level context
object throughout Printer.

This exposes a shortcoming of km::context::push/pop: to add an overlay
just for one rendering before returning a rendering result one has
to save that result to a temporary variable so that pop() could be
called. It would be great to have some convenience function that would
wrap a call km::mustache::render() into a push/pop pair. This is also
a matter of a separate commit.