Don't emit inlined main schemas
This optimisation skips emission of already inlined definitions while
retaining dependencies where they were. This may lead to confusing
file names (the file name will no more match the last schema defined
in it) but having the unused struct definition in the file is not very
helpful either. And in case with no dependencies, the whole file won't
be emitted anymore, which is nice.