Defer adding imports
This makes imports a bit more manageable - previously,
currentModel().addImportsFrom() was called along with analyzing
the model, making it difficult to step back and drop parts of the model
meant for substitution. This commit consolidates collection of imports
to be in two places: analyzeBodySchema() and resolveRef() - meaning
that model manipulations (not just additions, but removals too) can
freely occur before the final stages for a given API definition file.