Partition both allParams and vars by 'required'
Partitioning vars (allParams was partitioned before) became important
with libQuotient adding default initialisation of structure fields that
are not required. Syntactically those structures would be correct even
with the order of fields as before; but initialisation of those
structures in the code becomes tricky as one has to carefully
over-initialise even fields that have default values.