setList: add even empty lists
This allows to disable untamed propagation of values from higher levels.
At the same time, turns out, that propagation is almost unused in
the current templates for Quotient.

Fun fact: the code that enabled that propagation seems to have been
a somewhat premature optimisation introduced about a year into
the project existence, back in 2017.

Closes #50.