Complete initial support for local $refs
As of this commit, GTAD should be able to resolve $refs pointing to
a component in the same file (those starting with `#`). Non-local $refs
to fragments (that refer to another file and also have the # part) are
still not supported (Matrix API has none of those yet, anyway).

What could still be improved is ability to have a common component
defined once in the generated code and referred to from other parts of
the generated code. This is clearly impossible for things like
parameters but certainly possible for schemas.