Printer: don't enclose path literals in quotes any more
This kind of adornment can easily be done in Mustache if passed the
right data in the context. From now on, pathParts doesn't just include
a list of path parts for insertion but instead is formed as
`pathParts: [ literal|variable ]`, with `literal` and `variable` being
mutually exclusive and set to the value of the given path component only
when it is of the respective kind. In other words, a path like
"/room/{roomId}/events" will be translated into a sequence of 3 path
parts as follows:
```yaml
pathParts:
  literal: /room/
  variable: roomId
  literal: /events
```