Override descriptions for $refs more flexibly
Previously if a $ref has a sibling `description` it was automatically
inlined, assuming that because the description applies to this
individual type usage we have to have a dedicated definition for it.
That doesn't work very well though: Matrix 1.12 has got definitions in
capabilities.yaml that GTAD translated into type usages with seemingly
correct types and descriptions borrowed from the overriding
`description` attributes - but without the actual type definition.

This commit fixes that: when the analyzeSchema determines that the
$ref'ed schema has a name, it is imported (without applying an
overriding description, of course), while the overriding description is
still attached to the parameter/field corresponding to that schema, just
like until now.